(feat): 시각적 성취도 조회 기존 로직에 service grpc호출 추가

+++ constant/common/common.go
@@ -13,7 +13,7 @@ const (
 )
 
 // 수집률로 시각적 성취도 단계 분류
-func PercentCal(percentage uint) (Code string) {
+func PercentCal(percentage uint64) (Code string) {
 	switch {
 	case percentage >= 80:
 		Code = Ultimate
